Change MainMenuScene buttons to use pointers to commands
By declaring a bunch of commands first then assigning them later to a Button::onClickCommand(), the unique_ptr holds onto memory on the stack, which resulted in the error in the previous commit. Directly creating a new instance of it by resetting the pointer in MaineMenuScene.cpp solves the problem.