Improve sound load times
Load times were long since each instance of the PlaceShip and Attack
commands loaded their files over and over again. We use the SoundBuffers
found in Scene::Context instead to reuse the loaded resources. Ideally,
the game should be using resource holders so that there's no need to
pass the entire Context to the Attack command in order to swap between
different SoundBuffers.